Update unit tests for QueueActor
