+++ core/src/test/kotlin/com/malinskiy/marathon/execution/TestLogsProvider.kt
@@ -1,0 +1,15 @@
+package com.malinskiy.marathon.execution
+
+import com.malinskiy.marathon.report.logs.BatchLogs
+import com.malinskiy.marathon.report.logs.LogReport
+import com.malinskiy.marathon.report.logs.LogsProvider
+
+class TestLogsProvider(private val logs: Map<String, BatchLogs>) : LogsProvider {
+
+    override fun getFullReport(): LogReport {
+        return LogReport(logs)
+    }
+
+    override suspend fun getBatchReport(batchId: String): BatchLogs? = logs[batchId]
+
+}
